More serializer work; moved some methods out to mixins
