+++ backend/recipes/nested.py
@@ -1,11 +1,12 @@
 from rest_framework.serializers import ModelSerializer
 
+from recipes.mixins import GetImageMixin
 from recipes.models import Recipe
 
 
-class RecipeNestedSerializer(ModelSerializer):
+class RecipeNestedSerializer(ModelSerializer, GetImageMixin):
     """Сериализатор модели рецепта для вкладывания в другие сериализаторы"""
 
     class Meta:
         model = Recipe
-        fields = ('id', 'name', 'cooking_time')
+        fields = ('id', 'name', 'image', 'cooking_time')
